+++ jgclark.Dashboard/src/settingControllers.js
@@ -100,6 +100,34 @@ export async function toggleTomorrowSection(): Promise<void> {
 }
 
 /**
+ * Toggle showing the 'Quarter' section of the dashboard, by changing the setting file
+ * and then refreshing the dashboard.
+ */
+export async function toggleQuarterSection(): Promise<void> {
+  try {
+    // Get plugin settings
+    const config: dashboardConfigType = await getSettings()
+    // logDebug('toggleQuarterSection', `starting with existing value ${String(config.showQuarterSection)}`)
+
+    if (config == null || Object.keys(config).length === 0) {
+      throw new Error(
+        `Cannot find settings for the '${pluginID}' plugin. Please make sure you have installed it from the Plugin Preferences pane.`,
+      )
+    }
+    config.showQuarterSection = !config.showQuarterSection
+
+    // Save it back
+    const res = await saveSettings(pluginID, config)
+    // logDebug('toggleQuarterSection', `result -> ${String(res)}`)
+    logDebug('toggleQuarterSection', `-> new value ${String(config.showQuarterSection)}`)
+    logDebug('toggleQuarterSection', `------- now Refresh ---------`)
+    await showDashboard()
+  } catch (error) {
+    logError(pluginJson, `toggleQuarterSection: ${error.name}: ${error.message}`)
+  }
+}
+
+/**
  * Toggle showing the 'Week' section of the dashboard, by changing the setting file
  * and then refreshing the dashboard.
  */